pyproject.toml
[tool.poetry]
name = "tune-manager"
version = "0.1.0"
description = ""
authors = ["Evan Purkhiser <evanpurkhiser@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.5"
sqlalchemy = "^1.2"
watchdog = "^0.9.0"
mutagen = "^1.42"
requests = "^2.21"
pillow = "^5.4"
sanic = "^18.12"
sanic-cors = "^0.9.7"
beautifulsoup4 = "^4.7"
keyfinder = "^1.1"
sentry_sdk = "^0.13.1"

[tool.poetry.scripts]
tunemanager= 'tune_manager:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
